% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AfterBariatricSurgery}
\alias{AfterBariatricSurgery}
\title{RNAseq gene expression data for patients after Bariatric surgery}
\format{
{
RNAseq gene expression:
\describe{
\item{Gene_Name}{The name of the gene.}
\item{Gene_Expression}{The amount of expression of the gene shown.}
}
}
}
\source{
INSERM, Laurence Tiret (2019).
}
\usage{
AfterBariatricSurgery
}
\description{
RNAseq gene expression data for patients after Bariatric surgery
}
\examples{
\dontrun{
 AfterBariatricSurgery
}
}
\keyword{datasets}
