% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCpGIslands.R
\name{countCpGIslands}
\alias{countCpGIslands}
\title{Calculates the number of CpG islands in a DNA strand and creates a markdown file to highlight them}
\usage{
countCpGIslands(nucleotides)
}
\arguments{
\item{nucleotides}{A string of nucleotides in a DNA strand}
}
\value{
The number of CpG islands found and a file that shows where the islands are in the DNA strand
\itemize{
\item file - A string value for the file where the markdown text can be found
\item numIslands - An integer value for tthe amount of CpG islands found
}
}
\description{
A function that returns the ratio of CG nucleotides to the amount of all
nucleotides, and the ratio of observed:expected CpG Islands
}
\examples{
\dontrun{
countCpGIslands(PossibleCpGIslands)
}

}
