% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencesInGeneExpression.R
\name{differencesInGeneExpression}
\alias{differencesInGeneExpression}
\title{Bar Graph showing difference in gene expression of n genes}
\usage{
differencesInGeneExpression(RNAseq1, RNAseq2, n = 1)
}
\arguments{
\item{RNAseq1}{A dataframe of names of genes and their count values as
expression quantities from a RNAseq experiment}

\item{RNAseq2}{A dataframe of names of genes and their count values as
expression quantities from a RNAseq experiment}

\item{n}{The amount of genes with the largest difference in gene expression
to be displayed in the bar graph. Default value is 1.}
}
\value{
A bar graph indicating the amount of expression for the two RNAseq
datasets for the n genes of highest differences between the two datasets
}
\description{
A function that returns a bar graph showing gene expression of n genes
with greatest difference of expression between two RNA seq data inputs
}
\examples{
\dontrun{
differencesInGeneExpression(BeforeBariatricSurgery, AfterBariatricSurgery)
}

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York, 2016.
}
