% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCpGIslands.R
\name{findCpGIslands}
\alias{findCpGIslands}
\title{Highlighting all CpG Islands}
\usage{
findCpGIslands(nucleotides)
}
\arguments{
\item{nucleotides}{A string of nucleotides of a DNA sequence}
}
\value{
A visual output of highlighted CpG Islands in the DNA sequence
}
\description{
A function that returns a visual output highlighting all CpG islands found
in the DNA strand.
}
\examples{
\dontrun{
findCpGIslands(PossibleCpGIslands)
}


}
