% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencesInGeneExpression.R
\name{findDifferenceOfExpression}
\alias{findDifferenceOfExpression}
\title{Calculate a certain amount of genes with the largest difference in expression.}
\usage{
findDifferenceOfExpression(RNAseq1, RNAseq2, n)
}
\arguments{
\item{RNAseq1}{A dataframe of names of genes and their count values as
expression quantities from a RNAseq experiment}

\item{RNAseq2}{A dataframe of names of genes and their count values as
expression quantities from a RNAseq experiment}

\item{n}{The amount of genes with the largest difference in gene expression
to be displayed in the bar graph.}
}
\value{
A list with gene name and expression of the n genes with largest
difference in gene expression between the two RNAseq datasets.
\itemize{
\item GeneNames - The gene names with the highest difference in expression
\item Expressions - The RNAseq expression from each RNAseq dataset for
each gene in GeneNames
}
}
\description{
A helper function that returns n genes with the largest difference in expression
between two RNAseq datasets
}
\details{
This function is an internal function and therefore cannot be used directly
when installing this package - it is instead a helper function for
differencesInGeneExpression
}
