% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDifferencesInDNASequence.R
\name{FindDifferencesInDNASequence}
\alias{FindDifferencesInDNASequence}
\title{Highlighting differences in nucleotides between two DNA Sequences}
\usage{
FindDifferencesInDNASequence(strand1, strand2)
}
\arguments{
\item{strand1}{A string of nucleotides of one DNA sequence}

\item{strand2}{A string of nucleotides of another DNA sequence}
}
\value{
The html of the two DNA sequences, as well as a visual output of
the highlighted differences in nucleotides between two DNA sequences, as
well
}
\description{
A function that returns a visual output highlighting all differences in
nucleotides between two DNA sequences
}
\examples{
\dontrun{
FindDifferencesInDNASequence(BeforeBariatricSurgery, AfterBariatricSurgery)
}

}
\references{
Kevin Ushey, JJ Allaire, Hadley Wickham and Gary Ritchie (2020). rstudioapi:
Safely Access the RStudio API. R package version 0.13.
https://CRAN.R-project.org/package=rstudioapi
}
