% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MethylationObservation}
\alias{MethylationObservation}
\title{Two DNA sequences to see before and after bisulfite conversion}
\format{
{
Two DNA sequences:
\describe{
\item{Original DNA Sequence}{The DNA sequence before any modifications were made.}
\item{Post-Bisulfite Conversion DNA Sequence}{The DNA sequence after bisulfite conversion.}
}
}
}
\source{
fake source for now
}
\usage{
MethylationObservation
}
\description{
Two DNA sequences. One is the original sequence, the other is the sequence
after bisulfite conversion
}
\examples{
\dontrun{
 MethylationObservation
}
}
\keyword{datasets}
