% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCpGIslands.R
\name{observedAndExpected}
\alias{observedAndExpected}
\title{Calculate CG ratio and observed:expected ratio for CpG Islands}
\usage{
observedAndExpected(numC, numG, numCpG, lenNuc)
}
\arguments{
\item{numC}{The number of cytosines in a DNA strand}

\item{numG}{The number of guanines in a DNA strand}

\item{numCpG}{The number of cytosines followed by a guanine in a DNA strand}

\item{lenNuc}{The number of nucleotides in a DNA strand}
}
\value{
A dataframe with the CG ratio and observed:expected ratio of CpG Islands
\itemize{
\item CGRatio - A value for the CG ratio in a DNA strand
\item OERatio - A value for the observed:expected CpG Island ratio
}
}
\description{
A helper function that returns the ratio of CG nucleotides to the amount of all
nucleotides, and the ratio of observed:expected CpG Islands
}
\details{
This function is an internal function and therefore cannot be used directly
when installing this package - it is instead a helper function for
findCpGIslands
}
