% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PossibleCpGIslands}
\alias{PossibleCpGIslands}
\title{DNA sequence to Check For CpGIslands}
\format{
a string of lowercase characters that only include 'g', 'c', 't', 'a'
}
\source{
fake source for now
}
\usage{
PossibleCpGIslands
}
\description{
A string of nucleotides that will be used to check for CpGIslands
}
\examples{
\dontrun{
 PossibleCpGIslands
}
}
\keyword{datasets}
