% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMethylExpress.R
\name{runMethylExpress}
\alias{runMethylExpress}
\title{Launch Shiny App for MethylExpress}
\usage{
runMethylExpress()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for MethylExpress
}
\examples{
\dontrun{

MethylExpress::runMethylExpress()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
