% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BeforeBariatricSurgery}
\alias{BeforeBariatricSurgery}
\title{RNAseq gene expression data for patients before Bariatric surgery}
\format{
{
RNAseq gene expression:
\describe{
\item{Gene_Name}{The name of the gene.}
\item{Gene_Expression}{The amount of expression of the gene shown.}
}
}
}
\source{
INSERM, Laurence Tiret (2019).
}
\usage{
BeforeBariatricSurgery
}
\description{
RNAseq gene expression data for patients before Bariatric surgery
}
\examples{
\dontrun{
 BeforeBariatricSurgery
}
}
\references{
Poitou C, Perret C, Mathieu F, Truong V et al. Bariatric Surgery Induces
Disruption in Inflammatory Signaling Pathways Mediated by Immune Cells in
Adipose Tissue: A RNA-Seq Study. PLoS One 2015;10(5):e0125718.
PMID: 25938420
}
\keyword{datasets}
