% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCpGIslands.R
\name{countCpGIslands}
\alias{countCpGIslands}
\title{Calculates the number of CpG islands in a DNA strand and creates a markdown
file to highlight them}
\usage{
countCpGIslands(nucleotides)
}
\arguments{
\item{nucleotides}{A string of nucleotides in a DNA strand}
}
\value{
The number of CpG islands found and a file that shows where the islands are in the DNA strand
\itemize{
\item file - A string value for the file where the markdown text can be found
\item numIslands - An integer value for the amount of CpG islands found
}
}
\description{
A helper function that returns the file who contains markdown of the DNA
strand with the CpG islands highlighted, as well as the number of CpG
islands in the DNA strand
}
\details{
This function is an internal function and therefore cannot be used directly
when installing this package - it is instead a helper function for
findCpGIslands
}
