% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferencesInGeneExpression.R
\name{DifferencesInGeneExpression}
\alias{DifferencesInGeneExpression}
\title{Bar Graph showing difference in gene expression of n genes}
\usage{
DifferencesInGeneExpression(RNAseq1, RNAseq2, n = 1)
}
\arguments{
\item{RNAseq1}{A dataframe of names of genes and their expression quantities
from a RNAseq experiment}

\item{RNAseq2}{A dataframe of names of genes and their expression quantities
from a RNAseq experiment}

\item{n}{The amount of genes with the highest differences to be displayed in
in the bar graph. Default value is 1.}
}
\value{
A bar graph indicating the amount of expression for the two RNAseq
datasets for the n genes of highest differences between the two datasets
}
\description{
A function that returns a bar graph showing n genes of greatest difference
between two RNA seq data inputs
}
\examples{
\dontrun{
DifferencesInGeneExpression(RNAseq1, RNAseq2)
}

}
\references{
Reference 1 - DNA source
Reference 2 - highlight package
}
