% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDifferencesInDNASequence.R
\name{DNASequenceHighlights}
\alias{DNASequenceHighlights}
\title{Generate a markdown file with differing nucleotides highlighted}
\usage{
DNASequenceHighlights(strand1, strand2)
}
\arguments{
\item{strand1}{A string of nucleotides of one DNA sequence}

\item{strand2}{A string of nucleotides of another DNA sequence}
}
\value{
The html of the two DNA sequences (and highlight html)
}
\description{
A helper function that returns a file that includes highlight html to indicate
which nucleotides between two strands differ
}
\details{
This function is an internal function and therefore cannot be used directly
when installing this package - it is instead a helper function for
findDifferencesInDNASequence
}
