% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCpGIslands.R
\name{FindCpGIslands}
\alias{FindCpGIslands}
\title{Highlighting all CpG Islands}
\usage{
FindCpGIslands(nucleotides)
}
\arguments{
\item{nucleotides}{A string of nucleotides of a DNA sequence}
}
\value{
A visual output of highlighted CpG Islands in the DNA sequence
}
\description{
A function that returns a visual output highlighting all CpG islands found
in the DNA strand.
}
\examples{
\dontrun{
FindCpGIslands(nucleotides)
}

}
\references{
Reference 1 - DNA source
Reference 2 - highlight package
}
