% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCpGIslands.R
\name{findCpGIslands}
\alias{findCpGIslands}
\title{Highlighting all CpG Islands}
\usage{
findCpGIslands(nucleotides)
}
\arguments{
\item{nucleotides}{A string of nucleotides of a DNA sequence}
}
\value{
A visual output of highlighted CpG Islands in the DNA sequence
}
\description{
A function that returns a visual output highlighting all CpG islands found
in the DNA strand.
}
\details{
More details about how the CpG islands are calculated can be found in the
README and the vignette, but a quick overview as to how nucleotides are
considered part of a CpG islands is:
\enumerate{
\item the number of CGs must be > 50\% of the part of DNA being looked at
\item the observed number of CGs - to - expected number of CGs ratio must be > 60\%
}
}
\examples{
\dontrun{
findCpGIslands(PossibleCpGIslands)
}

}
\references{
Kevin Ushey, JJ Allaire, Hadley Wickham and Gary Ritchie (2020). rstudioapi:
Safely Access the RStudio API. R package version 0.13.
https://CRAN.R-project.org/package=rstudioapi
}
