% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCpGIslands.R
\name{findCpGIslands}
\alias{findCpGIslands}
\title{Highlighting all CpG Islands}
\usage{
findCpGIslands(nucleotides)
}
\arguments{
\item{nucleotides}{A string of nucleotides of a DNA sequence}
}
\value{
A visual output of highlighted CpG Islands in the DNA sequence
}
\description{
A function that returns a visual output highlighting all CpG islands found
in the DNA strand.
}
\examples{
\dontrun{
findCpGIslands(PossibleCpGIslands)
}

}
\references{
Kevin Ushey, JJ Allaire, Hadley Wickham and Gary Ritchie (2020). rstudioapi:
Safely Access the RStudio API. R package version 0.13.
https://CRAN.R-project.org/package=rstudioapi
}
