% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferencesInGeneExpression.R
\name{FindDifferenceOfExpression}
\alias{FindDifferenceOfExpression}
\title{Calculate a certain amount of genes with the largest difference in expression.}
\usage{
FindDifferenceOfExpression(RNAseq1, RNAseq2, n)
}
\arguments{
\item{RNAseq1}{A dataframe of names of genes and their expression quantities
from a RNAseq experiment}

\item{RNAseq2}{A dataframe of names of genes and their expression quantities
from a RNAseq experiment}

\item{n}{The amount of genes with the highest differences to be displayed in
in the bar graph.}
}
\value{
A list with the Gene name and expression of each RNAseq.
\itemize{
\item GeneNames - The gene names with the highest difference in expression
\item Expressions - The RNAseq expressions for each gene in GeneNames
}
}
\description{
A function that returns the nth amount of genes with the largest difference
in expression between the two RNAseq results.
}
\examples{
\dontrun{
FindDifferenceOfExpression(BeforeBariatricSurgery, AfterBariatricSurgery, 1)
}

}
