% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMethylatedCytosines.R
\name{findMatchingSections}
\alias{findMatchingSections}
\title{Generate a markdown file with all methylated cytosines highlighted}
\usage{
findMatchingSections(orig, bisulfite)
}
\arguments{
\item{orig}{A string of nucleotides of the original DNA sequence with no modifications}

\item{bisulfite}{A string of nucleotides of the DNA sequence after bisulfite conversion}
}
\value{
The file containing html of highlighted methylated nucleotides, as
well as a visual output of all nucleotides given, where the methylated
cytosines are highlighted.
}
\description{
A function that returns a markkdown file of the two input DNA strands, where
each cytosine that was methylated is highlighted
}
\examples{
\dontrun{
findMatchingSections(orig, bisulfite)
}

}
\references{
Reference 1 - DNA source
}
