% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMethylatedCytosines.R
\name{FindMethylatedCytosines}
\alias{FindMethylatedCytosines}
\title{Highlighting all Methylated Cytosines}
\usage{
FindMethylatedCytosines(orig, bisulfite)
}
\arguments{
\item{orig}{A string of nucleotides of the original DNA sequence with no modifications}

\item{bisulfite}{A string of nucleotides of the DNA sequence after bisulfite conversion}
}
\value{
A visual output of all nucleotides given, where the methylated cytosines
are highlighted.
}
\description{
A function that returns a visual output highlighting all methylated cytosines.
}
\examples{
\dontrun{
FindMethylatedCytosines(orig, bisulfite)
}

}
\references{
Reference 1 - DNA source
Reference 2 - highlight package
}
