% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Highlighting parts of text}
\usage{
highlight(fileName, nucleotides, highlighting)
}
\arguments{
\item{fileName}{The name of the file where the markdown will be saved to}

\item{nucleotides}{A string of nucleotides of a DNA sequence}

\item{highlighting}{A boolean to say if nucleotides will be highlighted in fileName}
}
\value{
A visual output of highlighted letters in a string
}
\description{
A function that returns a visual output highlighting specified letters in
a string.
}
\examples{
\dontrun{
highlight(fileName, nucleotides, TRUE)
}

}
