% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MethylationObservation}
\alias{MethylationObservation}
\title{Two DNA sequences to see before and after bisulfite conversion}
\format{
{
Two DNA sequences:
\describe{
\item{originalDNA}{The DNA sequence before any modifications were made.}
\item{bisulfite}{The DNA sequence after bisulfite conversion.}
}
}
}
\source{
University of Georgia, Hunt BG, Rehan S (2018).
}
\usage{
MethylationObservation
}
\description{
Two DNA sequences. One is the original sequence, the other is the sequence
after bisulfite conversion
}
\examples{
\dontrun{
 MethylationObservation
}
}
\references{
Hunt, B. G. (2018). GEO Accession viewer. Retrieved November 17, 2020,
from https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE111611
}
\keyword{datasets}
